requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
platformdirs==2.5.4
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
virtualenv==20.16.7
